timeout: use uint64_t for all plain time values
This enforces the use of unsigned 64-bit values for time in the duration type,
the timeout framework, the timer session, the userland timer-drivers, and the
alarm framework on all platforms. The commit also adapts the code that uses
these tools accross all basic repositories (base, base-*, os. gems, libports,
ports, dde_*) to use unsigned 64-bit values for time as well as far as this
does not imply profound modifications.

Fixes #3208

+++ repos/base-hw/src/test/cpu_quota/main.cc
@@ -130,7 +130,7 @@ struct Main
 		log("start measurement ...");
 		timer.sigh(timer_signal.cap);
 
-		auto measure = [&] (unsigned duration_sec) {
+		auto measure = [&] (uint64_t duration_sec) {
 			timer.trigger_once(duration_sec * 1000 * 1000);
 			synchronizer.synchronize();
 			timer_signal.receive();
