+++ repos/base/src/lib/timeout/timeout.cc
@@ -48,7 +48,7 @@ void Timeout::discard()
  ** Timeout::Alarm **
  ********************/
 
-bool Timeout::Alarm::_on_alarm(unsigned)
+bool Timeout::Alarm::_on_alarm(uint64_t)
 {
 	if (handler) {
 		Handler *current = handler;
@@ -68,7 +68,7 @@ Timeout::Alarm::~Alarm()
 }
 
 
-bool Timeout::Alarm::Raw::is_pending_at(unsigned long time, bool time_period) const
+bool Timeout::Alarm::Raw::is_pending_at(uint64_t time, bool time_period) const
 {
 	return (time_period == deadline_period &&
 	        time        >= deadline) ||
@@ -83,12 +83,12 @@ bool Timeout::Alarm::Raw::is_pending_at(unsigned long time, bool time_period) co
 
 void Alarm_timeout_scheduler::handle_timeout(Duration duration)
 {
-	unsigned long const curr_time_us = duration.trunc_to_plain_us().value;
+	uint64_t const curr_time_us = duration.trunc_to_plain_us().value;
 
 	_alarm_handle(curr_time_us);
 
 	/* sleep time is either until the next deadline or the maximum timout */
-	unsigned long sleep_time_us;
+	uint64_t sleep_time_us;
 	Alarm::Time deadline_us;
 	if (_alarm_next_deadline(&deadline_us)) {
 		sleep_time_us = deadline_us - curr_time_us;
@@ -139,7 +139,7 @@ void Alarm_timeout_scheduler::_schedule_one_shot(Timeout      &timeout,
                                                  Microseconds  duration)
 {
 	/* raise timeout duration by the age of the local time value */
-	unsigned long us = _time_source.curr_time().trunc_to_plain_us().value;
+	uint64_t us = _time_source.curr_time().trunc_to_plain_us().value;
 	if (us >= _now) {
 		us = duration.value + (us - _now); }
 	else {
@@ -294,7 +294,7 @@ void Alarm_timeout_scheduler::_alarm_handle(Alarm::Time curr_time)
 		_pending_head = _pending_head->_next;
 		curr->_next = nullptr;
 
-		unsigned long triggered = 1;
+		uint64_t triggered = 1;
 
 		if (curr->_raw.period) {
 			Alarm::Time deadline = curr->_raw.deadline;
