+++ repos/os/src/drivers/audio/spec/linux/main.cc
@@ -188,7 +188,7 @@ class Audio_out::Out
 		{
 			_timer.sigh(_timer_dispatcher);
 
-			unsigned const us = (Audio_out::PERIOD * 1000 / Audio_out::SAMPLE_RATE)*1000;
+			uint64_t const us = (Audio_out::PERIOD * 1000 / Audio_out::SAMPLE_RATE)*1000;
 			_timer.trigger_periodic(us);
 		}
