+++ repos/os/src/drivers/framebuffer/spec/exynos5/driver.cc
@@ -36,7 +36,7 @@ class Timer_delayer : public Mmio::Delayer, public Timer::Connection
 
 		Timer_delayer(Genode::Env &env) : Timer::Connection(env) { }
 
-		void usleep(unsigned us) { Timer::Connection::usleep(us); }
+		void usleep(uint64_t us) { Timer::Connection::usleep(us); }
 };
