+++ repos/os/src/drivers/sd_card/spec/rpi/driver.h
@@ -152,7 +152,7 @@ class Sd_card::Driver : public  Driver_base,
 		{
 			Timer_delayer(Genode::Env &env) : Timer::Connection(env) { }
 
-			void usleep(unsigned us) override { Timer::Connection::usleep(us); }
+			void usleep(uint64_t us) override { Timer::Connection::usleep(us); }
 		};
 
 		Env            &_env;
