+++ repos/os/src/init/heartbeat.h
@@ -34,7 +34,7 @@ class Init::Heartbeat : Genode::Noncopyable
 
 		Constructible<Timer::Connection> _timer { };
 
-		unsigned _rate_ms = 0;
+		uint64_t _rate_ms = 0;
 
 		Signal_handler<Heartbeat> _timer_handler;
