+++ repos/os/src/test/fb_bench/main.cc
@@ -46,7 +46,7 @@ struct Test
 		memset(buf[1], ~0, fb_ds.size());
 	}
 
-	void conclusion(unsigned kib, unsigned start_ms, unsigned end_ms) {
+	void conclusion(unsigned kib, uint64_t start_ms, uint64_t end_ms) {
 		log("throughput: ", kib / (end_ms - start_ms), " MiB/sec"); }
 
 	~Test() { log("\nTEST ", id, " finished\n"); }
@@ -67,7 +67,7 @@ struct Bytewise_ram_test : Test
 	Bytewise_ram_test(Env &env, int id) : Test(env, id, brief)
 	{
 		unsigned       kib      = 0;
-		unsigned const start_ms = timer.elapsed_ms();
+		uint64_t const start_ms = timer.elapsed_ms();
 		for (; timer.elapsed_ms() - start_ms < DURATION_MS;) {
 			memcpy(buf[0], buf[1], fb_ds.size());
 			kib += fb_ds.size() / 1024;
@@ -83,7 +83,7 @@ struct Bytewise_fb_test : Test
 	Bytewise_fb_test(Env &env, int id) : Test(env, id, brief)
 	{
 		unsigned       kib      = 0;
-		unsigned const start_ms = timer.elapsed_ms();
+		uint64_t const start_ms = timer.elapsed_ms();
 		for (unsigned i = 0; timer.elapsed_ms() - start_ms < DURATION_MS; i++) {
 			memcpy(fb_ds.local_addr<char>(), buf[i % 2], fb_ds.size());
 			kib += fb_ds.size() / 1024;
@@ -99,7 +99,7 @@ struct Blit_test : Test
 	Blit_test(Env &env, int id) : Test(env, id, brief)
 	{
 		unsigned       kib      = 0;
-		unsigned const start_ms = timer.elapsed_ms();
+		uint64_t const start_ms = timer.elapsed_ms();
 		unsigned const w        = fb_mode.width() * fb_mode.bytes_per_pixel();
 		unsigned const h        = fb_mode.height();
 		for (unsigned i = 0; timer.elapsed_ms() - start_ms < DURATION_MS; i++) {
@@ -117,7 +117,7 @@ struct Unaligned_blit_test : Test
 	Unaligned_blit_test(Env &env, int id) : Test(env, id, brief)
 	{
 		unsigned       kib      = 0;
-		unsigned const start_ms = timer.elapsed_ms();
+		uint64_t const start_ms = timer.elapsed_ms();
 		unsigned const w        = fb_mode.width() * fb_mode.bytes_per_pixel();
 		unsigned const h        = fb_mode.height();
 		for (unsigned i = 0; timer.elapsed_ms() - start_ms < DURATION_MS; i++) {
